move seed to sample app and add waffle gen
